fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@22675 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/RepositoryBrowser.cpp
@@ -4345,7 +4345,7 @@ void CRepositoryBrowser::OnNMCustomdrawRepolist(NMHDR *pNMHDR, LRESULT *pResult)
         // Tell Windows to paint the control itself.
         *pResult = CDRF_DODEFAULT;
 
-        if (m_RepoList.GetItemCount() > pLVCD->nmcd.dwItemSpec)
+        if (m_RepoList.GetItemCount() > (int)pLVCD->nmcd.dwItemSpec)
         {
             COLORREF crText = GetSysColor(COLOR_WINDOWTEXT);
             CAutoReadLock locker(m_guard);
